Fix a Wformat warning.
git-svn-id: https://llvm.org/svn/llvm-project/klee/trunk@102877 91177308-0d34-0410-b5e6-96231b3b80d8

+++ lib/Core/Executor.cpp
@@ -555,7 +555,7 @@ void Executor::initializeGlobals(ExecutionState &state) {
         if (end && *end == '\0') {
           klee_message("NOTE: allocated global at asm specified address: %#08llx"
                        " (%llu bytes)",
-                       address, size);
+                       (long long) address, (unsigned long long) size);
           mo = memory->allocateFixed(address, size, &*i);
           mo->isUserSpecified = true; // XXX hack;
         }
